Fix Binary/Text serialization in omitnull
This could be an unpopular change as omitnull previously could decode
any text by virtue of it being "" or a value. And it would decode that
as either set or unset, but there was never a capability for null.

Presumably if you are using omitnull in a scenario where you require
this behavior it's likely a mistake since the states representable here
with only empty string and string (for example) are the same as omit and
you should just use that.

- Add a omit|null|set byte to the front of text serialization for
  omitnull
- Fix a bug where BinaryMarshaler never returned []byte{1} properly